Refactor job files API toward micro-service.
Refactor the API controller into a plain Python class with JSON inputs/outputs that mimic a generic rest interface without any framework stuff cruft (a similar pattern is used lib/galaxy/tool_util/deps/views.py). This should allow building a simple Flask micro service (or whatever else) from the same code.
