+++ lib/galaxy/job_execution/files/__init__.py
@@ -1,0 +1,3 @@
+from .view import JobFilesView
+
+__all__ = ('JobFilesView', )
