+++ lib/galaxy/webapps/galaxy/api/job_files.py
@@ -1,22 +1,12 @@
 """ API for asynchronous job running mechanisms can use to fetch or put files
 related to running and queued jobs.
 """
-import logging
-import os
-import shutil
-
-from galaxy import (
-    exceptions,
-    model,
-    util
-)
+from galaxy.job_execution.files import JobFilesView
 from galaxy.web import (
     expose_api_anonymous_and_sessionless,
     expose_api_raw_anonymous_and_sessionless,
 )
 from galaxy.webapps.base.controller import BaseAPIController
-
-log = logging.getLogger(__name__)
 
 
 class JobFilesAPIController(BaseAPIController):
@@ -30,6 +20,9 @@ class JobFilesAPIController(BaseAPIController):
     low-level unfiltered files and such authorization would break Galaxy's
     security model for tool execution.
     """
+
+    def __init__(self, app):
+        self._job_files_view = JobFilesView(app)
 
     @expose_api_raw_anonymous_and_sessionless
     def index(self, trans, job_id, **kwargs):
@@ -53,9 +46,7 @@ class JobFilesAPIController(BaseAPIController):
         :rtype:     binary
         :returns:   contents of file
         """
-        self.__authorize_job_access(trans, job_id, **kwargs)
-        path = kwargs.get("path", None)
-        return open(path, 'rb')
+        return self._job_files_view.get_file(job_id, **kwargs)
 
     @expose_api_anonymous_and_sessionless
     def create(self, trans, job_id, payload, **kwargs):
@@ -81,84 +72,5 @@ class JobFilesAPIController(BaseAPIController):
         :rtype:     dict
         :returns:   an okay message
         """
-        job = self.__authorize_job_access(trans, job_id, **payload)
-        path = payload.get("path")
-        self.__check_job_can_write_to_path(trans, job, path)
-
-        # Is this writing an unneeded file? Should this just copy in Python?
-        if '__file_path' in payload:
-            file_path = payload.get('__file_path')
-            upload_store = trans.app.config.nginx_upload_job_files_store
-            assert upload_store, ("Request appears to have been processed by"
-                                  " nginx_upload_module but Galaxy is not"
-                                  " configured to recognize it")
-            assert file_path.startswith(upload_store), \
-                ("Filename provided by nginx (%s) is not in correct"
-                 " directory (%s)" % (file_path, upload_store))
-            input_file = open(file_path)
-        else:
-            input_file = payload.get("file",
-                                     payload.get("__file", None)).file
-        target_dir = os.path.dirname(path)
-        util.safe_makedirs(target_dir)
-        try:
-            shutil.move(input_file.name, path)
-        finally:
-            try:
-                input_file.close()
-            except OSError:
-                # Fails to close file if not using nginx upload because the
-                # tempfile has moved and Python wants to delete it.
-                pass
-        return {"message": "ok"}
-
-    def __authorize_job_access(self, trans, encoded_job_id, **kwargs):
-        for key in ["path", "job_key"]:
-            if key not in kwargs:
-                error_message = "Job files action requires a valid '%s'." % key
-                raise exceptions.ObjectAttributeMissingException(error_message)
-
-        job_id = trans.security.decode_id(encoded_job_id)
-        job_key = trans.security.encode_id(job_id, kind="jobs_files")
-        if not util.safe_str_cmp(kwargs["job_key"], job_key):
-            raise exceptions.ItemAccessibilityException("Invalid job_key supplied.")
-
-        # Verify job is active. Don't update the contents of complete jobs.
-        job = trans.sa_session.query(model.Job).get(job_id)
-        if job.finished:
-            error_message = "Attempting to read or modify the files of a job that has already completed."
-            raise exceptions.ItemAccessibilityException(error_message)
-        return job
-
-    def __check_job_can_write_to_path(self, trans, job, path):
-        """ Verify an idealized job runner should actually be able to write to
-        the specified path - it must be a dataset output, a dataset "extra
-        file", or a some place in the working directory of this job.
-
-        Would like similar checks for reading the unstructured nature of loc
-        files make this very difficult. (See abandoned work here
-        https://gist.github.com/jmchilton/9103619.)
-        """
-        in_work_dir = self.__in_working_directory(job, path, trans.app)
-        if not in_work_dir and not self.__is_output_dataset_path(job, path):
-            raise exceptions.ItemAccessibilityException("Job is not authorized to write to supplied path.")
-
-    def __is_output_dataset_path(self, job, path):
-        """ Check if is an output path for this job or a file in the an
-        output's extra files path.
-        """
-        da_lists = [job.output_datasets, job.output_library_datasets]
-        for da_list in da_lists:
-            for job_dataset_association in da_list:
-                dataset = job_dataset_association.dataset
-                if not dataset:
-                    continue
-                if os.path.abspath(dataset.file_name) == os.path.abspath(path):
-                    return True
-                elif util.in_directory(path, dataset.extra_files_path):
-                    return True
-        return False
-
-    def __in_working_directory(self, job, path, app):
-        working_directory = app.object_store.get_filename(job, base_dir='job_work', dir_only=True, extra_dir=str(job.id))
-        return util.in_directory(path, working_directory)
+        payload.update(kwargs)
+        return self._job_files_view.post_file(job_id, **payload)
